End java boost project

+++ app/src/main/java/com/dzyuba/javaboost/domain/entities/lesson/LessonItem.kt
@@ -1,6 +1,6 @@
 package com.dzyuba.javaboost.domain.entities.lesson
 
 abstract class LessonItem(
-    open val id: Int,
+    open val id: Long,
     open val type: Type
 )
